requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
css-html-js-minify==2.5.5
cssselect==1.2.0
cssutils==2.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.1
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-mptt==0.14.0
django-otp==1.1.6
django-otp-twilio==1.0.2
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
freetype-py==2.3.0
frozenlist==1.3.3
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.4
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
pandas==2.0.1
Pillow==9.4.0
premailer==3.10.0
pycairo==1.23.0
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
reportlab==4.0.0
reportlib==3.4.0
requests==2.28.2
rlPyCairo==0.2.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
twilio==8.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.9.1
